Add Bcrypt (notes)
This is a good exercise to try once. by adding the Gem I learned how bcrypt stores a string of `password_digest` through virtual attributes of `password` and `password_confirmation`. Some of my questions are how to clean up my tests because I had to get a little janky with passwords inside specs. I know that's not the right way to do it. Bcrypt uses a the method `has_secure_password` in User. 

This has also been a good task to get me comfortable with a workflow pace of keeping the tests green. Rather than focus on fast and features, I'm focused on working through the tedious parts. The good news is I've now learned these specifics in a way that I'm going to be able to remember them! :-)